requirements.txt
tensorflow>=2.8
numpy
matplotlib
opencv-python
pillow
